% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-landing.R
\name{make_description}
\alias{make_description}
\title{Make Description Text}
\usage{
make_description(rawHTML = FALSE, ...)
}
\arguments{
\item{rawHTML}{Logical: FALSE by default. If TRUE, raw HTML string should be
supplied. FALSE and {shiny} functions should be utilized to create text.}

\item{...}{Character string of HTML or {shiny} functions}
}
\value{
HTML for description text in landing page.
}
\description{
Make Description Text
}
\examples{


if (interactive()) {

# To use {shiny} functions

make_description(rawHTML = FALSE,
tagList(
 column(3, p( "You have been selected to participate in a task that does...")),
 column(3, p("This task will take approximately ... it is optional..."))
)
)

# To use HTML

make_description(rawHTML = TRUE,
'
     <div class = "col-sm-3 box mx-4 p-4 text-center">
       You have been selected to participate in a task that does...
     </div>
     <div class = "col-sm-3 box mx-4 p-4 text-center">
      <center>
         <p>This task will take approximately ... it is optional...</p>
       </center>
     </div>
     <div class = "col-sm-3 box mx-4 p-4 text-center">
      <center>
         <p>You will be compensated ... </p>
       </center>
     </div>
 ')

}


}
