% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-landing.R
\name{make_informed_consent}
\alias{make_informed_consent}
\title{Make Informed Consent Text}
\usage{
make_informed_consent(rawHTML = FALSE, ...)
}
\arguments{
\item{rawHTML}{Logical: FALSE by default. If TRUE, raw HTML string should be
supplied. FALSE and {shiny} functions should be utilized to create text.}

\item{...}{Character string of HTML or {shiny} functions}
}
\value{
HTML for informed consent text in landing page.
}
\description{
Make Informed Consent Text
}
\seealso{
make_description
}
