% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{shinyland}
\alias{shinyland}
\title{Implement a landing page}
\usage{
shinyland(
  navbar_title,
  task_title,
  description_text,
  informed_consent_text,
  task_url
)
}
\arguments{
\item{navbar_title}{The title for the Navigation Bar}

\item{task_title}{The task title}

\item{description_text}{Task description text -- see \code{\link{make_description}}}

\item{informed_consent_text}{Informed consent language -- see \code{\link{make_informed_consent}}}

\item{task_url}{The URL for the task where a user will be sent upon agreeing to participate.}
}
\value{
HTML, CSS, and JavaScript necessary to create a task's landing page.
}
\description{
Implement a landing page
}
